+++ src/collective/contact/core/tests/test_adapters.py
@@ -1,11 +1,17 @@
 # -*- coding: utf8 -*-
-
+import datetime
 import unittest2 as unittest
+
+from zope.intid.interfaces import IIntIds
+from zope.component import getUtility
+from plone import api
+from plone.app.testing.interfaces import TEST_USER_NAME
+from z3c.relationfield.relation import RelationValue
 
 from ecreall.helpers.testing.base import BaseTest
 
 from collective.contact.core.testing import INTEGRATION
-from collective.contact.core.interfaces import IVCard
+from collective.contact.core.interfaces import IVCard, IPersonHeldPositions
 
 
 class TestAdapters(unittest.TestCase, BaseTest):
@@ -54,3 +60,44 @@ class TestAdapters(unittest.TestCase, BaseTest):
         self.assertEqual(vcard.fn.value, u'Régiment H'.encode('utf-8'))
         self.assertEqual(vcard.kind.value, 'org')
         self.assertTrue(hasattr(vcard, 'adr'))
+
+    def test_person_held_positions(self):
+        self.login(TEST_USER_NAME)
+        degaulle = self.directory.degaulle
+        intids = getUtility(IIntIds)
+        # de gaulle in 1960...
+        api.content.create(container=self.directory, type='organization', id='france',
+                           )
+        api.content.create(container=degaulle, type='held_position', id='president',
+                           start_date=datetime.date(1959,1,8),
+                           position=RelationValue(intids.getId(self.directory.france)))
+        api.content.create(container=degaulle, type='held_position', id='lieutenant-colonel',
+                           position=RelationValue(intids.getId(self.directory.armeedeterre)),
+                           end_date=datetime.date(1940,6,1),
+                           start_date=datetime.date(1933,12,25))
+        api.content.create(container=degaulle, type='held_position', id='commandant',
+                           position=RelationValue(intids.getId(self.directory.armeedeterre)),
+                           start_date=datetime.date(1927,10,9),
+                           end_date=datetime.date(1933,12,25))
+        del self.degaulle.gadt.end_date
+        self.degaulle.moveObjectsToTop(['president'])
+
+        adapter = IPersonHeldPositions(degaulle)
+        self.assertEqual(adapter.get_main_position(), self.degaulle.president)
+        self.assertEqual(adapter.get_current_positions(),
+                         (self.degaulle.president,
+                          self.degaulle.gadt, ))
+        self.assertEqual(adapter.get_sorted_positions(),
+                         (self.degaulle.president,
+                          self.degaulle.gadt,
+                          degaulle.adt,
+                          degaulle['lieutenant-colonel'],
+                          self.degaulle['commandant'],
+                          ))
+
+        api.content.transition(degaulle.adt, 'deactivate')
+        self.assertEqual(adapter.get_main_position(), self.degaulle.president)
+
+        api.content.transition(degaulle.president, 'deactivate')
+        self.assertEqual(adapter.get_main_position(), degaulle.gadt)
+